+++ elfi/methods/utils.py
@@ -1,14 +1,13 @@
 """This module contains utilities for methods."""
 
 import logging
+from functools import partial
 from math import ceil
+from typing import Union
 
+import matplotlib.pyplot as plt
 import numpy as np
 import scipy.stats as ss
-
-import elfi.model.augmenter as augmenter
-from elfi.clients.native import Client
-from elfi.model.elfi_model import ComputationContext
 
 logger = logging.getLogger(__name__)
 
@@ -91,6 +90,23 @@ def normalize_weights(weights):
     return w / wsum
 
 
+def compute_ess(weights: Union[None, np.ndarray] = None):
+    """Compute the Effective Sample Size (ESS). Weights are assumed to be unnormalized.
+
+    Parameters
+    ----------
+    weights: unnormalized weights
+
+    """
+    # normalize weights
+    weights = normalize_weights(weights)
+
+    # compute ESS
+    numer = np.square(np.sum(weights))
+    denom = np.sum(np.square(weights))
+    return numer / denom
+
+
 def weighted_var(x, weights=None):
     """Unbiased weighted variance (sample variance) for the components of x.
 
@@ -117,10 +133,10 @@ def weighted_var(x, weights=None):
         weights = np.ones(len(x))
 
     V_1 = np.sum(weights)
-    V_2 = np.sum(weights**2)
+    V_2 = np.sum(weights ** 2)
 
     xbar = np.average(x, weights=weights, axis=0)
-    numerator = weights.dot((x - xbar)**2)
+    numerator = weights.dot((x - xbar) ** 2)
     s2 = numerator / (V_1 - (V_2 / V_1))
     return s2
 
@@ -216,7 +232,6 @@ class GMDistribution:
         n_accepted = 0
         n_left = size
         trials = 0
-
         while n_accepted < size:
             inds = random_state.choice(len(means), size=n_left, p=weights)
             rvs = means[inds]
@@ -301,127 +316,6 @@ def numgrad(fn, x, h=None, replace_neg_inf=True):
     return grad[1, :]
 
 
-# TODO: check that there are no latent variables in parameter parents.
-#       pdfs and gradients wouldn't be correct in those cases as it would require
-#       integrating out those latent variables. This is equivalent to that all
-#       stochastic nodes are parameters.
-# TODO: could use some optimization
-class ModelPrior:
-    """Construct a joint prior distribution over all the parameter nodes in `ElfiModel`."""
-
-    def __init__(self, model):
-        """Initialize a ModelPrior.
-
-        Parameters
-        ----------
-        model : ElfiModel
-
-        """
-        model = model.copy()
-        self.parameter_names = model.parameter_names
-        self.dim = len(self.parameter_names)
-        self.client = Client()
-
-        # Prepare nets for the pdf methods
-        self._pdf_node = augmenter.add_pdf_nodes(model, log=False)[0]
-        self._logpdf_node = augmenter.add_pdf_nodes(model, log=True)[0]
-
-        self._rvs_net = self.client.compile(model.source_net, outputs=self.parameter_names)
-        self._pdf_net = self.client.compile(model.source_net, outputs=self._pdf_node)
-        self._logpdf_net = self.client.compile(model.source_net, outputs=self._logpdf_node)
-
-    def rvs(self, size=None, random_state=None):
-        """Sample the joint prior."""
-        random_state = np.random if random_state is None else random_state
-
-        context = ComputationContext(size or 1, seed='global')
-        loaded_net = self.client.load_data(self._rvs_net, context, batch_index=0)
-
-        # Change to the correct random_state instance
-        # TODO: allow passing random_state to ComputationContext seed
-        loaded_net.nodes['_random_state'].update({'output': random_state})
-        del loaded_net.nodes['_random_state']['operation']
-
-        batch = self.client.compute(loaded_net)
-        rvs = np.column_stack([batch[p] for p in self.parameter_names])
-
-        if self.dim == 1:
-            rvs = rvs.reshape(size or 1)
-
-        return rvs[0] if size is None else rvs
-
-    def pdf(self, x):
-        """Return the density of the joint prior at x."""
-        return self._evaluate_pdf(x)
-
-    def logpdf(self, x):
-        """Return the log density of the joint prior at x."""
-        return self._evaluate_pdf(x, log=True)
-
-    def _evaluate_pdf(self, x, log=False):
-        if log:
-            net = self._logpdf_net
-            node = self._logpdf_node
-        else:
-            net = self._pdf_net
-            node = self._pdf_node
-
-        x = np.asanyarray(x)
-        ndim = x.ndim
-        x = x.reshape((-1, self.dim))
-        batch = self._to_batch(x)
-
-        # TODO: we could add a seed value that would load a "random state" instance
-        #       throwing an error if it is used, for instance seed="not used".
-        context = ComputationContext(len(x), seed=0)
-        loaded_net = self.client.load_data(net, context, batch_index=0)
-
-        # Override
-        for k, v in batch.items():
-            loaded_net.nodes[k].update({'output': v})
-            del loaded_net.nodes[k]['operation']
-
-        val = self.client.compute(loaded_net)[node]
-        if ndim == 0 or (ndim == 1 and self.dim > 1):
-            val = val[0]
-
-        return val
-
-    def gradient_pdf(self, x):
-        """Return the gradient of density of the joint prior at x."""
-        raise NotImplementedError
-
-    def gradient_logpdf(self, x, stepsize=None):
-        """Return the gradient of log density of the joint prior at x.
-
-        Parameters
-        ----------
-        x : float or np.ndarray
-        stepsize : float or list
-            Stepsize or stepsizes for the dimensions
-
-        """
-        x = np.asanyarray(x)
-        ndim = x.ndim
-        x = x.reshape((-1, self.dim))
-
-        grads = np.zeros_like(x)
-
-        for i in range(len(grads)):
-            xi = x[i]
-            grads[i] = numgrad(self.logpdf, xi, h=stepsize)
-
-        grads[np.isinf(grads)] = 0
-        grads[np.isnan(grads)] = 0
-
-        if ndim == 0 or (ndim == 1 and self.dim > 1):
-            grads = grads[0]
-        return grads
-
-    def _to_batch(self, x):
-        return {p: x[:, i] for i, p in enumerate(self.parameter_names)}
-
-
 def sample_object_to_dict(data, elem, skip=''):
     """Process data from self object to data dictionary to prepare for json serialization.
 
@@ -482,3 +376,435 @@ def numpy_to_python_type(data):
                 data[key] = int(val)
             elif 'float' in data_type:
                 data[key] = float(val)
+
+
+def weighted_sample_quantile(x, alpha, weights=None):
+    """Calculate alpha-quantile of a weighted sample.
+
+    Parameters
+    ----------
+    x : array
+        One-dimensional sample
+    alpha : float
+        Probability threshold for alpha-quantile
+    weights : array, optional
+        Sample weights (possibly unnormalized), equal weights by default
+
+    Returns
+    -------
+    alpha_q : array
+        alpha-quantile
+
+    """
+    index = np.argsort(x)
+    if alpha == 0:
+        alpha_q = x[index[0]]
+    else:
+        if weights is None:
+            weights = np.ones(len(index))
+        weights = weights / np.sum(weights)
+        sorted_weights = weights[index]
+        cum_weights = np.insert(np.cumsum(sorted_weights), 0, 0)
+        cum_weights[-1] = 1.0
+        index_alpha = np.where(np.logical_and(cum_weights[:-1] < alpha,
+                                              alpha <= cum_weights[1:]))[0][0]
+        alpha_q = x[index][index_alpha]
+
+    return alpha_q
+
+
+class DensityRatioEstimation:
+    """A density ratio estimation class."""
+
+    def __init__(self,
+                 n=100,
+                 epsilon=0.1,
+                 max_iter=500,
+                 abs_tol=0.01,
+                 conv_check_interval=20,
+                 fold=5,
+                 optimize=False):
+        """Construct the density ratio estimation algorithm object.
+
+        Parameters
+        ----------
+        n : int
+            Number of RBF basis functions.
+        epsilon : float
+            Parameter determining speed of gradient descent.
+        max_iter : int
+            Maximum number of iterations used in gradient descent optimization of the weights.
+        abs_tol : float
+            Absolute tolerance value for determining convergence of optimization of the weights.
+        conv_check_interval : int
+            Integer defining the interval of convergence checks in gradient descent.
+        fold : int
+            Number of folds in likelihood cross validation used to optimize basis scale-params.
+        optimize : boolean
+            Boolean indicating whether or not to optimize RBF scale.
+
+        """
+        self.n = n
+        self.epsilon = epsilon
+        self.max_iter = max_iter
+        self.abs_tol = abs_tol
+        self.fold = fold
+        self.sigma = None
+        self.conv_check_interval = conv_check_interval
+        self.optimize = optimize
+
+    def fit(self,
+            x,
+            y,
+            weights_x=None,
+            weights_y=None,
+            sigma=None):
+        """Fit the density ratio estimation object.
+
+        Parameters
+        ----------
+        x : array
+            Sample from the nominator distribution.
+        y : sample
+            Sample from the denominator distribution.
+        weights_x : array
+            Vector of non-negative nominator sample weights, must be able to normalize.
+        weights_y : array
+            Vector of non-negative denominator sample weights, must be able to normalize.
+        sigma : float or list
+            List of RBF kernel scales, fit selected at initial call.
+
+        """
+        self.x_len = x.shape[0]
+        self.y_len = y.shape[0]
+        x = x.reshape(self.x_len, -1)
+        y = y.reshape(self.y_len, -1)
+        self.x = x
+
+        if self.x_len < self.n:
+            raise ValueError("Number of RBFs ({}) can't be larger "
+                             "than number of samples ({}).".format(self.n, self.x_len))
+
+        self.theta = x[:self.n, :]
+        if weights_x is None:
+            weights_x = np.ones(self.x_len)
+        if weights_y is None:
+            weights_y = np.ones(self.y_len)
+
+        self.weights_x = weights_x / np.sum(weights_x)
+        self.weights_y = weights_y / np.sum(weights_y)
+
+        self.x0 = np.average(x, axis=0, weights=weights_x)
+
+        if isinstance(sigma, float):
+            self.sigma = sigma
+            self.optimize = False
+        if self.optimize:
+            if isinstance(sigma, list):
+                scores_tuple = zip(*[self._KLIEP_lcv(x, y, sigma_i)
+                                   for sigma_i in sigma])
+
+                self.sigma = sigma[np.argmax(scores_tuple)]
+            else:
+                raise ValueError("To optimize RBF scale, "
+                                 "you need to provide a list of candidate scales.")
+
+        if self.sigma is None:
+            raise ValueError("RBF width (sigma) has to provided in first call.")
+
+        A = self._compute_A(x, self.sigma)
+        b, b_normalized = self._compute_b(y, self.sigma)
+
+        alpha = self._KLIEP(A, b, b_normalized, weights_x, self.sigma)
+        self.w = partial(self._weighted_basis_sum, sigma=self.sigma, alpha=alpha)
+
+    def _gaussian_basis(self, x, x0, sigma):
+        """N-D RBF basis-function with equal scale-parameter for every dim."""
+        return np.exp(-0.5 * np.sum((x - x0) ** 2) / sigma / sigma)
+
+    def _weighted_basis_sum(self, x, sigma, alpha):
+        """Weighted sum of gaussian basis functions evaluated at x."""
+        return np.dot(np.array([[self._gaussian_basis(j, i, sigma) for j in self.theta]
+                                for i in np.atleast_2d(x)]), alpha)
+
+    def _compute_A(self, x, sigma):
+        A = np.array([[self._gaussian_basis(i, j, sigma) for j in self.theta] for i in x])
+        return A
+
+    def _compute_b(self, y, sigma):
+        b = np.sum(np.array(
+                [[self._gaussian_basis(i, y[j, :], sigma) * self.weights_y[j]
+                  for j in np.arange(self.y_len)]
+                 for i in self.theta]), axis=1)
+        b_normalized = b / np.dot(b.T, b)
+        return b, b_normalized
+
+    def _KLIEP_lcv(self, x, y, sigma):
+        """Compute KLIEP scores for fold-folds."""
+        A = self._compute_A(x, sigma)
+        b, b_normalized = self._compute_b(y, sigma)
+
+        non_null = np.any(A > 1e-64, axis=1)
+        non_null_length = sum(non_null)
+        if non_null_length == 0:
+            return np.Inf
+
+        A_full = A[non_null, :]
+        x_full = x[non_null, :]
+        weights_x_full = self.weights_x[non_null]
+
+        fold_indices = np.array_split(np.arange(non_null_length), self.fold)
+        score = np.zeros(self.fold)
+        for i_fold, fold_index in enumerate(fold_indices):
+            fold_index_minus = np.setdiff1d(np.arange(non_null_length), fold_index)
+            alpha = self._KLIEP(A=A_full[fold_index_minus, :], b=b, b_normalized=b_normalized,
+                                weights_x=weights_x_full[fold_index_minus], sigma=sigma)
+            score[i_fold] = np.average(
+                np.log(self._weighted_basis_sum(x_full[fold_index, :], sigma, alpha)),
+                weights=weights_x_full[fold_index])
+
+        return [np.mean(score)]
+
+    def _KLIEP(self, A, b, b_normalized, weights_x, sigma):
+        """Kullback-Leibler Importance Estimation Procedure using gradient descent."""
+        alpha = 1 / self.n * np.ones(self.n)
+        target_fun_prev = self._weighted_basis_sum(x=self.x, sigma=sigma, alpha=alpha)
+        abs_diff = 0.0
+        non_null = np.any(A > 1e-64, axis=1)
+        A_full = A[non_null, :]
+        weights_x_full = weights_x[non_null]
+        for i in np.arange(self.max_iter):
+            dAdalpha = np.matmul(A_full.T, (weights_x_full / (np.matmul(A_full, alpha))))
+            alpha += self.epsilon * dAdalpha
+            alpha = np.maximum(0, alpha + (1 - np.dot(b.T, alpha)) * b_normalized)
+            alpha = alpha / np.dot(b.T, alpha)
+            if np.remainder(i, self.conv_check_interval) == 0:
+                target_fun = self._weighted_basis_sum(x=self.x, sigma=sigma, alpha=alpha)
+                abs_diff = np.linalg.norm(target_fun - target_fun_prev)
+                if abs_diff < self.abs_tol:
+                    break
+                target_fun_prev = target_fun
+        return alpha
+
+    def max_ratio(self):
+        """Find the maximum of the density ratio at numerator sample."""
+        max_value = np.max(self.w(self.x))
+        return max_value
+
+
+class NDimBoundingBox:
+    """Class for the n-dimensional bounding box built around the optimal point."""
+
+    def __init__(self, rotation, center, limits, eps_region):
+        """Class initialiser.
+
+        Parameters
+        ----------
+        rotation: (D,D) rotation matrix for the Bounding Box
+        center: (D,) center of the Bounding Box
+        limits: np.ndarray, shape: (D,2)
+            The limits of the bounding box.
+
+        """
+        assert rotation.ndim == 2
+        assert center.ndim == 1
+        assert limits.ndim == 2
+        assert limits.shape[1] == 2
+        assert center.shape[0] == rotation.shape[0] == rotation.shape[1]
+
+        self.rotation = rotation
+        self.center = center
+        self.limits = limits
+        self.dim = rotation.shape[0]
+        self.eps_region = eps_region
+
+        self.rotation_inv = np.linalg.inv(self.rotation)
+
+        self.volume = self._compute_volume()
+
+    def _compute_volume(self):
+        v = np.prod(- self.limits[:, 0] + self.limits[:, 1])
+
+        if v == 0:
+            logger.warning("zero volume area")
+            v = 0.05
+        return v
+
+    def contains(self, point):
+        """Check if point is inside the bounding box.
+
+        Parameters
+        ----------
+        point: (D, )
+
+        Returns
+        -------
+        True/False
+
+        """
+        assert point.ndim == 1
+        assert point.shape[0] == self.dim
+
+        # transform to bb coordinate system
+        point1 = np.dot(self.rotation_inv, point) + np.dot(self.rotation_inv, -self.center)
+
+        # Check if point is inside bounding box
+        inside = True
+        for i in range(point1.shape[0]):
+            if (point1[i] < self.limits[i][0]) or (point1[i] > self.limits[i][1]):
+                inside = False
+                break
+        return inside
+
+    def sample(self, n2, seed=None):
+        """Sample n2 points from the posterior.
+
+        Parameters
+        ----------
+        n2: int
+        seed: seed of the sampling procedure
+
+        Returns
+        -------
+        np.ndarray, shape: (n2,D)
+
+        """
+        center = self.center
+        limits = self.limits
+        rot = self.rotation
+
+        loc = limits[:, 0]
+        scale = limits[:, 1] - limits[:, 0]
+
+        # draw n2 samples
+        theta = []
+        for i in range(loc.shape[0]):
+            rv = ss.uniform(loc=loc[i], scale=scale[i])
+            theta.append(rv.rvs(size=(n2, 1), random_state=seed))
+
+        theta = np.concatenate(theta, -1)
+        # translate and rotate
+        theta_new = np.dot(rot, theta.T).T + center
+
+        return theta_new
+
+    def pdf(self, theta: np.ndarray):
+        """Evalute the pdf.
+
+        Parameters
+        ----------
+        theta: np.ndarray (D,)
+
+        Returns
+        -------
+        float
+
+        """
+        return self.contains(theta) / self.volume
+
+    def plot(self, samples):
+        """Plot the bounding box (works only if dim=1 or dim=2).
+
+        Parameters
+        ----------
+        samples: np.ndarray, shape: (N, D)
+
+        Returns
+        -------
+        None
+
+        """
+        R = self.rotation
+        T = self.center
+        lim = self.limits
+
+        if self.dim == 1:
+            plt.figure()
+            plt.title("Bounding Box region")
+
+            # plot eigenectors
+            end_point = T + R[0, 0] * lim[0][0]
+            plt.plot([T[0], end_point[0]], [T[1], end_point[1]], "r-o")
+            end_point = T + R[0, 0] * lim[0][1]
+            plt.plot([T[0], end_point[0]], [T[1], end_point[1]], "r-o")
+
+            plt.plot(samples, np.zeros_like(samples), "bo")
+            plt.legend()
+            plt.show(block=False)
+        else:
+            plt.figure()
+            plt.title("Bounding Box region")
+
+            # plot sampled points
+            plt.plot(samples[:, 0], samples[:, 1], "bo", label="samples")
+
+            # plot eigenectors
+            x = T
+            x1 = T + R[:, 0] * lim[0][0]
+            plt.plot([T[0], x1[0]], [T[1], x1[1]], "y-o", label="-v1")
+            x3 = T + R[:, 0] * lim[0][1]
+            plt.plot([T[0], x3[0]], [T[1], x3[1]], "g-o", label="v1")
+
+            x2 = T + R[:, 1] * lim[1][0]
+            plt.plot([T[0], x2[0]], [T[1], x2[1]], "k-o", label="-v2")
+            x4 = T + R[:, 1] * lim[1][1]
+            plt.plot([T[0], x4[0]], [T[1], x4[1]], "c-o", label="v2")
+
+            # plot boundaries
+            def plot_side(x, x1, x2):
+                tmp = x + (x1 - x) + (x2 - x)
+                plt.plot([x1[0], tmp[0], x2[0]], [x1[1], tmp[1], x2[1]], "r-o")
+
+            plot_side(x, x1, x2)
+            plot_side(x, x2, x3)
+            plot_side(x, x3, x4)
+            plot_side(x, x4, x1)
+
+            plt.legend()
+            plt.show(block=False)
+
+
+def flat_array_to_dict(names, arr):
+    """Map flat array to a dictionary with parameter names.
+
+    Parameters
+    ----------
+    names: List[string]
+        parameter names
+    arr: np.array, shape: (D,)
+        flat theta array
+
+    Returns
+    -------
+    Dict
+       dictionary with named parameters
+
+    """
+    # res = model.generate(batch_size=1)
+    # param_dict = {}
+    # cur_ind = 0
+    # for param_name in model.parameter_names:
+    #     tensor = res[param_name]
+    #     assert isinstance(tensor, np.ndarray)
+    #     if tensor.ndim == 2:
+    #         dim = tensor.shape[1]
+    #         val = arr[cur_ind:cur_ind + dim]
+    #         cur_ind += dim
+    #         assert isinstance(val, np.ndarray)
+    #         assert val.ndim == 1
+    #         param_dict[param_name] = np.expand_dims(val, 0)
+    #
+    #     else:
+    #         dim = 1
+    #         val = arr[cur_ind:cur_ind + dim]
+    #         cur_ind += dim
+    #         assert isinstance(val, np.ndarray)
+    #         assert val.ndim == 1
+    #         param_dict[param_name] = val
+
+    # TODO: This approach covers only the case where all parameters
+    # TODO: are univariate variables (i.e. independent between them)
+    param_dict = {}
+    for ii, param_name in enumerate(names):
+        param_dict[param_name] = np.expand_dims(arr[ii:ii + 1], 0)
+    return param_dict